All repositories created, user - matchs - fields cr(u)d

+++ src/main/java/com/cactus/RolandGarrosEE/repositories/implementation/ArbitratorPersistentBean.java
@@ -20,7 +20,7 @@ public class ArbitratorPersistentBean implements ArbitratorPersistentRemote {
     public void deleteArbitrator(Arbitrator arbitrator) {
     }
 
-    public Optional<Arbitrator> getArbitratorById(long arbitratorId) {
+    public Optional<Arbitrator> getArbitratorById(int arbitratorId) {
         return Optional.empty();
     }
