+++ src/main/java/com/cactus/RolandGarrosEE/repositories/remotes/TournamentPersistentRemote.java
@@ -10,6 +10,6 @@ import java.util.Optional;
 public interface TournamentPersistentRemote {
     void saveTournament(Tournament tournament);
     void deleteTournament(Tournament tournament);
-    Optional<Tournament> getTournamentById(long tournamentId);
+    Optional<Tournament> getTournamentById(int tournamentId);
     List<Tournament> allTournament();
 }
